Move NGFF axis metadata to src/ngff/
