+++ src/dimension.h
@@ -5,25 +5,6 @@
 #include <stddef.h>
 #include <stdint.h>
 
-enum dimension_axis_type
-{
-  dimension_axis_space = 0, // default (zero-init = space)
-  dimension_axis_time,
-  dimension_axis_channel,
-  dimension_axis_other,
-};
-
-// OME-NGFF v0.5 axis metadata: unit, scale, and type.
-// Only consumed by zarr metadata writers.
-struct omengff_axis
-{
-  const char* unit; // axis unit (e.g. "micrometer"),
-                    // NULL defaults to "index" in metadata
-  double scale;     // physical pixel scale for coordinateTransformations
-                    // (must be non-negative; 0 treated as 1.0)
-  enum dimension_axis_type type; // space, time, channel, or other
-};
-
 struct dimension
 {
   uint64_t size; // 0 means unbounded (dim 0 only: stream indefinitely)
@@ -35,7 +16,6 @@ struct dimension
   uint8_t storage_position;  // position in storage layout (0=outermost).
                              // dims[0].storage_position must be 0.
                              // Must be a valid permutation of 0..rank-1.
-  struct omengff_axis ngff;  // OME-NGFF v0.5 axis metadata
 };
 
 // Initialize dims from a name string and sizes array.
