+++ src/ngff/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_src_lib(ngff_metadata SOURCES ngff_metadata.c ngff_metadata.h ngff_axis.h
+    LINKS json_writer dimension chucky_log
+)
