+++ src/ngff/ngff_metadata.c
@@ -1,0 +1,147 @@
+#include "ngff/ngff_metadata.h"
+#include "defs.limits.h"
+#include "zarr/json_writer.h"
+
+#include <stdio.h>
+
+int
+ngff_multiscale_group_json(char* buf,
+                           size_t cap,
+                           uint8_t rank,
+                           int nlod,
+                           const struct dimension* const* level_dims,
+                           const struct ngff_axis* axes)
+{
+  struct json_writer jw;
+  jw_init(&jw, buf, cap);
+
+  const struct dimension* l0 = level_dims[0];
+
+  jw_object_begin(&jw);
+
+  jw_key(&jw, "zarr_format");
+  jw_int(&jw, 3);
+
+  jw_key(&jw, "node_type");
+  jw_string(&jw, "group");
+
+  jw_key(&jw, "consolidated_metadata");
+  jw_null(&jw);
+
+  jw_key(&jw, "attributes");
+  jw_object_begin(&jw);
+
+  jw_key(&jw, "ome");
+  jw_object_begin(&jw);
+  jw_key(&jw, "version");
+  jw_string(&jw, "0.5");
+
+  jw_key(&jw, "multiscales");
+  jw_array_begin(&jw);
+  jw_object_begin(&jw);
+
+  jw_key(&jw, "axes");
+  jw_array_begin(&jw);
+  for (int d = 0; d < rank; ++d) {
+    jw_object_begin(&jw);
+    jw_key(&jw, "name");
+    if (l0[d].name)
+      jw_string(&jw, l0[d].name);
+    else {
+      char name[8];
+      snprintf(name, sizeof(name), "d%d", d);
+      jw_string(&jw, name);
+    }
+    jw_key(&jw, "type");
+    {
+      const char* type = "space";
+      if (axes) {
+        switch (axes[d].type) {
+          case ngff_axis_time:
+            type = "time";
+            break;
+          case ngff_axis_channel:
+            type = "channel";
+            break;
+          case ngff_axis_other:
+            type = "custom";
+            break;
+          default:
+            type = "space";
+            break;
+        }
+      }
+      jw_string(&jw, type);
+    }
+    if (axes && axes[d].unit) {
+      jw_key(&jw, "unit");
+      jw_string(&jw, axes[d].unit);
+    }
+    jw_object_end(&jw);
+  }
+  jw_array_end(&jw);
+
+  jw_key(&jw, "datasets");
+  jw_array_begin(&jw);
+  for (int lv = 0; lv < nlod; ++lv) {
+    jw_object_begin(&jw);
+    jw_key(&jw, "path");
+    char lvstr[8];
+    snprintf(lvstr, sizeof(lvstr), "%d", lv);
+    jw_string(&jw, lvstr);
+
+    // Precompute per-axis physical scale and downsample factor.
+    // Use the actual ratio of L0 size to level size so that dimensions
+    // which cannot be downsampled further (e.g. size 1) get factor 1.
+    const struct dimension* lv_dims = level_dims[lv];
+    double scale[MAX_ZARR_RANK], translation[MAX_ZARR_RANK];
+    for (int d = 0; d < rank; ++d) {
+      double phys = (axes && axes[d].scale > 0) ? axes[d].scale : 1.0;
+      double factor = 1.0;
+      uint64_t l0_size = l0[d].size;
+      uint64_t lv_size = lv_dims[d].size;
+      if (l0_size > 0 && lv_size > 0)
+        factor = (double)l0_size / (double)lv_size;
+      scale[d] = phys * factor;
+      translation[d] = 0.5 * phys * (factor - 1.0);
+    }
+
+    jw_key(&jw, "coordinateTransformations");
+    jw_array_begin(&jw);
+    // scale
+    jw_object_begin(&jw);
+    jw_key(&jw, "type");
+    jw_string(&jw, "scale");
+    jw_key(&jw, "scale");
+    jw_array_begin(&jw);
+    for (int d = 0; d < rank; ++d)
+      jw_float(&jw, scale[d]);
+    jw_array_end(&jw);
+    jw_object_end(&jw);
+    // translation
+    jw_object_begin(&jw);
+    jw_key(&jw, "type");
+    jw_string(&jw, "translation");
+    jw_key(&jw, "translation");
+    jw_array_begin(&jw);
+    for (int d = 0; d < rank; ++d)
+      jw_float(&jw, translation[d]);
+    jw_array_end(&jw);
+    jw_object_end(&jw);
+    jw_array_end(&jw);
+
+    jw_object_end(&jw);
+  }
+  jw_array_end(&jw);
+
+  jw_object_end(&jw); // multiscales[0]
+  jw_array_end(&jw);  // multiscales
+
+  jw_object_end(&jw); // ome
+  jw_object_end(&jw); // attributes
+  jw_object_end(&jw); // root
+
+  if (jw_error(&jw))
+    return -1;
+  return (int)jw_length(&jw);
+}
