+++ src/zarr/CMakeLists.txt
@@ -9,7 +9,7 @@ add_src_lib(shard_delivery SOURCES shard_delivery.c shard_delivery.h
 )
 
 add_src_lib(zarr_fs_sink SOURCES zarr_fs_sink.c ../zarr_fs_sink.h
-    LINKS zarr_metadata writer io_queue platform_io lod_plan dimension chucky_log
+    LINKS zarr_metadata ngff_metadata writer io_queue platform_io lod_plan dimension chucky_log
 )
 
 add_src_lib(s3_client SOURCES s3_client.c s3_client.h
@@ -17,5 +17,5 @@ add_src_lib(s3_client SOURCES s3_client.c s3_client.h
 )
 
 add_src_lib(zarr_s3_sink SOURCES zarr_s3_sink.c ../zarr_s3_sink.h
-    LINKS zarr_metadata s3_client writer lod_plan dimension chucky_log
+    LINKS zarr_metadata ngff_metadata s3_client writer lod_plan dimension chucky_log
 )
