+++ src/zarr_s3_sink.h
@@ -2,6 +2,7 @@
 
 #include "dimension.h"
 #include "dtype.h"
+#include "ngff/ngff_axis.h"
 #include "types.codec.h"
 #include "writer.h"
 
@@ -77,12 +78,13 @@ struct zarr_s3_multiscale_config
   const struct dimension* dimensions; // L0 dimensions
   int nlod;                           // number of levels (0 = auto)
   struct codec_config codec;
-  size_t part_size;          // 0 = default (8 MiB)
-  double throughput_gbps;    // 0 = default (10.0)
-  size_t max_retries;        // 0 = CRT default (10)
-  uint32_t backoff_scale_ms; // 0 = CRT default (500)
-  uint32_t max_backoff_secs; // 0 = CRT default (20)
-  uint64_t timeout_ns;       // 0 = no timeout (infinite)
+  const struct ngff_axis* axes; // per-dim NGFF axis metadata, NULL = defaults
+  size_t part_size;             // 0 = default (8 MiB)
+  double throughput_gbps;       // 0 = default (10.0)
+  size_t max_retries;           // 0 = CRT default (10)
+  uint32_t backoff_scale_ms;    // 0 = CRT default (500)
+  uint32_t max_backoff_secs;    // 0 = CRT default (20)
+  uint64_t timeout_ns;          // 0 = no timeout (infinite)
 };
 
 struct zarr_s3_multiscale_sink;
