+++ tests/CMakeLists.txt
@@ -113,7 +113,7 @@ add_gpu_test_exe(test_stream          test_stream.c          LINKS CUDA::cuda_dr
 add_gpu_test_exe(test_compress        test_compress.c        LINKS CUDA::cuda_driver CUDA::cudart_static stream Zstd::Zstd)
 add_gpu_test_exe(test_aggregate       test_aggregate.c       LINKS CUDA::cuda_driver CUDA::cudart_static aggregate index_ops_util)
 
-add_test_exe(test_json_writer test_json_writer.c zarr_metadata chucky_log)
+add_test_exe(test_json_writer test_json_writer.c zarr_metadata ngff_metadata chucky_log)
 add_test_exe(test_io_queue    test_io_queue.c    io_queue chucky_log)
 
 add_gpu_test_exe(test_zarr_fs_sink   test_zarr_fs_sink.c   LINKS CUDA::cuda_driver CUDA::cudart_static zarr_fs_sink stream Zstd::Zstd test_platform test_shard_verify test_voxel_encode)
